requirements.txt
kafka-python
flask
flask_restful
gevent
Werkzeug
requests
pymongo
